+++ examples/inline/example.js
@@ -1,0 +1,4 @@
+'use strict';
+require('./main.css');
+
+console.log('Hello World');