+++ examples/inline/webpack.config.js
@@ -1,0 +1,26 @@
+var HtmlWebpackPlugin = require('../..');
+var ExtractTextPlugin = require('extract-text-webpack-plugin');
+module.exports = {
+  entry: './example.js',
+  output: {
+    path: __dirname + '/dist',
+    publicPath: '',
+    filename: 'bundle.js'
+  },
+  module: {
+    loaders: [
+      { test: /\.css$/, loader: ExtractTextPlugin.extract('style-loader', 'css-loader') },
+      { test: /\.jade$/, loader: 'jade' }
+    ]
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      inject: false,
+      template: 'template.jade',
+      filename: 'index.html',
+      favicon: 'favicon.ico',
+      title: 'Jade demo'
+    }),
+    new ExtractTextPlugin('styles.css')
+  ]
+};